[io-layer] Remove Sleep and SleepEx
