+++ mono/io-layer/wthreads.c
@@ -30,7 +30,6 @@
 #include <mono/utils/mono-threads.h>
 #include <mono/utils/atomic.h>
 #include <mono/utils/mono-mutex.h>
-#include <mono/utils/mono-lazy-init.h>
 #include <mono/utils/mono-time.h>
 
 #ifdef HAVE_VALGRIND_MEMCHECK_H
@@ -197,145 +196,6 @@ gpointer
 wapi_get_current_thread_handle (void)
 {
 	return get_current_thread_handle ();
-}
-
-static mono_lazy_init_t sleepex_init = MONO_LAZY_INIT_STATUS_NOT_INITIALIZED;
-static mono_mutex_t sleepex_mutex;
-static mono_cond_t sleepex_cond;
-
-static void
-sleepex_initialize (void)
-{
-	mono_mutex_init (&sleepex_mutex);
-	mono_cond_init (&sleepex_cond, NULL);
-}
-
-static void
-sleepex_interrupt (gpointer data)
-{
-	mono_mutex_lock (&sleepex_mutex);
-	mono_cond_broadcast (&sleepex_cond);
-	mono_mutex_unlock (&sleepex_mutex);
-}
-
-static inline guint32
-sleepex_interruptable (guint32 ms)
-{
-	gboolean interrupted;
-	guint32 start, now, end;
-
-	g_assert (INFINITE == G_MAXUINT32);
-
-	start = mono_msec_ticks ();
-
-	if (start < G_MAXUINT32 - ms) {
-		end = start + ms;
-	} else {
-		/* start + ms would overflow guint32 */
-		end = G_MAXUINT32;
-	}
-
-	mono_lazy_initialize (&sleepex_init, sleepex_initialize);
-
-	mono_mutex_lock (&sleepex_mutex);
-
-	for (now = mono_msec_ticks (); ms == INFINITE || now - start < ms; now = mono_msec_ticks ()) {
-		mono_thread_info_install_interrupt (sleepex_interrupt, NULL, &interrupted);
-		if (interrupted) {
-			mono_mutex_unlock (&sleepex_mutex);
-			return WAIT_IO_COMPLETION;
-		}
-
-		if (ms < INFINITE)
-			mono_cond_timedwait_ms (&sleepex_cond, &sleepex_mutex, end - now);
-		else
-			mono_cond_wait (&sleepex_cond, &sleepex_mutex);
-
-		mono_thread_info_uninstall_interrupt (&interrupted);
-		if (interrupted) {
-			mono_mutex_unlock (&sleepex_mutex);
-			return WAIT_IO_COMPLETION;
-		}
-	}
-
-	mono_mutex_unlock (&sleepex_mutex);
-
-	return 0;
-}
-
-/**
- * SleepEx:
- * @ms: The time in milliseconds to suspend for
- * @alertable: if TRUE, the wait can be interrupted by an APC call
- *
- * Suspends execution of the current thread for @ms milliseconds.  A
- * value of zero causes the thread to relinquish its time slice.  A
- * value of %INFINITE causes an infinite delay.
- */
-guint32
-SleepEx (guint32 ms, gboolean alertable)
-{
-	if (ms == 0) {
-		MonoThreadInfo *info;
-
-		mono_thread_info_yield ();
-
-		info = mono_thread_info_current ();
-		if (info && mono_thread_info_is_interrupt_state (info))
-			return WAIT_IO_COMPLETION;
-
-		return 0;
-	}
-
-	if (alertable)
-		return sleepex_interruptable (ms);
-
-	DEBUG("%s: Sleeping for %d ms", __func__, ms);
-
-	if (ms == INFINITE) {
-		do {
-			sleep (G_MAXUINT32);
-		} while (1);
-	} else {
-		int ret;
-#if defined (__linux__) && !defined(PLATFORM_ANDROID)
-		struct timespec start, target;
-
-		/* Use clock_nanosleep () to prevent time drifting problems when nanosleep () is interrupted by signals */
-		ret = clock_gettime (CLOCK_MONOTONIC, &start);
-		g_assert (ret == 0);
-
-		target = start;
-		target.tv_sec += ms / 1000;
-		target.tv_nsec += (ms % 1000) * 1000000;
-		if (target.tv_nsec > 999999999) {
-			target.tv_nsec -= 999999999;
-			target.tv_sec ++;
-		}
-
-		do {
-			ret = clock_nanosleep (CLOCK_MONOTONIC, TIMER_ABSTIME, &target, NULL);
-		} while (ret != 0);
-#else
-		struct timespec req, rem;
-
-		req.tv_sec = ms / 1000;
-		req.tv_nsec = (ms % 1000) * 1000000;
-
-		do {
-			memset (&rem, 0, sizeof (rem));
-			ret = nanosleep (&req, &rem);
-		} while (ret != 0);
-#endif /* __linux__ */
-	}
-
-	return 0;
-}
-
-void
-Sleep(guint32 ms)
-{
-	SleepEx(ms, FALSE);
 }
 
 gboolean
